cornsnake/requirements.txt
aiohttp-devtools
aiohttp_debugtoolbar
psycopg2-binary
aiopg
SQLAlchemy
